attack phase and fortification test save file

+++ RiskGameProject/src/riskgame/model/Utils/InfoRetriver.java
@@ -324,7 +324,7 @@ public class InfoRetriver {
         for (int i = 0; i < playersList.size(); i++) {
             Player curPlayer = playersList.get(i);
             String curPlayerName = curPlayer.getPlayerName();
-            float curRatio = playerDomiViewObserver.getControlRatioList().get(i);
+            double curRatio = playerDomiViewObserver.getControlRatioList().get(i);
             dataList.add(new PieChart.Data(curPlayerName, curRatio));
 
             System.out.println("name:" + curPlayerName + ", ratio: " + curRatio);
